ThorlabsLTStage

ThorlabsLTStage.PS_3D

InstrumentConfig.initialize

ThorlabsLTStage.get_limits

ThorlabsLTStage.get_limits_x

ThorlabsLTStage.get_limits_y

ThorlabsLTStage.get_limits_z

ThorlabsLTStage.home

ThorlabsLTStage.home_x

ThorlabsLTStage.home_y

ThorlabsLTStage.home_z

ThorlabsLTStage.move_x_abs

ThorlabsLTStage.move_x_rel

ThorlabsLTStage.move_xyz

ThorlabsLTStage.move_y_abs

ThorlabsLTStage.move_y_rel

ThorlabsLTStage.move_z_abs

ThorlabsLTStage.move_z_rel

ThorlabsLTStage.pos_x

ThorlabsLTStage.pos_xyz

ThorlabsLTStage.pos_y

ThorlabsLTStage.pos_z

ThorlabsLTStage.set_limits

ThorlabsLTStage.PS_3D — Type

Available functions

move_xyz(xyz, x, y, z)

move_x_abs(xyz, x)

move_y_abs(xyz, y)

move_z_abs(xyz, z)

move_x_rel(xyz, x)

move_y_rel(xyz, y)

move_z_rel(xyz, z)

get_pos(xyz)

get_pos_x(xyz)

get_pos_y(xyz)

get_pos_z(xyz)

home(xyz)

home_x(xyz)

home_y(xyz)

home_z(xyz)

set_limits(xyz, low, high)

get_limits(xyz)

get_limits_x(xyz)

get_limits_y(xyz)

get_limits_z(xyz)

reset_limits(xyz)

InstrumentConfig.initialize — Method

xyz = initialize(ThorlabsLTS150)

Connect to Thorlabs device

Returns:

ThorlabsLTS150: Device Handle

ThorlabsLTStage.get_limits — Method

Returns

(lts.x_low_limit, lts.y_low_limit, lts.z_low_limit),
   (lts.x_high_limit, lts.y_high_limit, lts.z_high_limit)

ThorlabsLTStage.get_limits_x — Method

Returns

(x_low_limit, x_high_limit)

ThorlabsLTStage.get_limits_y — Method

Returns

(y_low_limit, y_high_limit)

ThorlabsLTStage.get_limits_z — Method

Returns

(z_low_limit, z_high_limit)

ThorlabsLTStage.home — Method

home(xyz)

This will home the x, y and z stages at the same time

ThorlabsLTStage.home_x — Method

home_x(xyz)

This will home the x stage

ThorlabsLTStage.home_y — Method

home_y(xyz)

This will home the y stage

ThorlabsLTStage.home_z — Method

home_z(xyz)

This will home the z stage

ThorlabsLTStage.move_x_abs — Method

move_x(xyz, x)

Moves x stage to desired absolute location

ThorlabsLTStage.move_x_rel — Method

move_x(xyz, x)

Moves x stage forward or backwards.

A positive number will move it forwards along the x axis while a negative number will move it backwards.

ThorlabsLTStage.move_xyz — Method

move_xyz(xyz, x, y, z)

Simulatenously moves x, y and z stage to desired location

ThorlabsLTStage.move_y_abs — Method

move_y(xyz, y)

Moves y stage to desired absolute location

ThorlabsLTStage.move_y_rel — Method

move_y(xyz, y)

Moves y stage forward or backwards.

A positive number will move it forwards along the y axis while a negative number will move it backwards.

ThorlabsLTStage.move_z_abs — Method

move_z(xyz, z)

Moves z stage to desired absolute location

ThorlabsLTStage.move_z_rel — Method

move_z(xyz, z)

Moves z stage forward or backwards.

A positive number will move it forwards along the z axis while a negative number will move it backwards.

ThorlabsLTStage.pos_x — Method

pos_x(xyz)

Returns the current position of x stage

ThorlabsLTStage.pos_xyz — Method

pos_xyz(xyz)

Returns location of x, y and z stages in the form of a Array: [x, y, z]

ThorlabsLTStage.pos_y — Method

pos_y(xyz)

Returns the current position of y stage

ThorlabsLTStage.pos_z — Method

pos_z(xyz)

Returns the current position of z stage

ThorlabsLTStage.set_limits — Method

set_limits(xyz, (x_low_lim, y_low_lim, z_low_lim), (x_high_lim, y_high_lim, z_high_lim))

Arguments

low: A Pair or an Array of three positions: (lts.x_low_limit, lts.y_low_limit, lts.z_low_limit)

high: A Pair or an Array of three positions: (lts.x_high_limit, lts.y_high_limit, lts.z_high_limit)