ThorlabsLTStage

ThorlabsLTStage.LTS_3D

ThorlabsLTStage.ThorlabsLTS150

InstrumentConfig.initialize

ThorlabsLTStage.get_limits

ThorlabsLTStage.get_limits_x

ThorlabsLTStage.get_limits_y

ThorlabsLTStage.get_limits_z

ThorlabsLTStage.home

ThorlabsLTStage.home_x

ThorlabsLTStage.home_y

ThorlabsLTStage.home_z

ThorlabsLTStage.microsteps_per_m

ThorlabsLTStage.microsteps_per_m

ThorlabsLTStage.microsteps_per_mm

ThorlabsLTStage.move_x_abs

ThorlabsLTStage.move_x_rel

ThorlabsLTStage.move_xyz

ThorlabsLTStage.move_y_abs

ThorlabsLTStage.move_y_rel

ThorlabsLTStage.move_z_abs

ThorlabsLTStage.move_z_rel

ThorlabsLTStage.pos_x

ThorlabsLTStage.pos_xyz

ThorlabsLTStage.pos_y

ThorlabsLTStage.pos_z

ThorlabsLTStage.set_limits

ThorlabsLTStage.LTS_3D — Type

Available functions

move_xyz(xyz, x, y, z)

pos_xyz(xyz, x, y, z)

move_x_abs(xyz, x)

move_y_abs(xyz, y)

move_z_abs(xyz, z)

move_x_rel(xyz, x)

move_y_rel(xyz, y)

move_z_rel(xyz, z)

pos_xyz(xyz)

pos_x(xyz)

pos_y(xyz)

pos_z(xyz)

home(xyz)

home_x(xyz)

home_y(xyz)

home_z(xyz)

set_limits(xyz, low, high)

get_limits(xyz)

get_limits_x(xyz)

get_limits_y(xyz)

get_limits_z(xyz)

reset_limits(xyz)

ThorlabsLTStage.ThorlabsLTS150 — Type

Available Functions

initialize

move_xyz(xyz, x, y, z)

pos_xyz(xyz, x, y, z)

move_x_abs(xyz, x)

move_y_abs(xyz, y)

move_z_abs(xyz, z)

move_x_rel(xyz, x)

move_y_rel(xyz, y)

move_z_rel(xyz, z)

pos_xyz(xyz)

pos_x(xyz)

pos_y(xyz)

pos_z(xyz)

home(xyz)

home_x(xyz)

home_y(xyz)

home_z(xyz)

set_limits(xyz, low, high)

get_limits(xyz)

get_limits_x(xyz)

get_limits_y(xyz)

get_limits_z(xyz)

clear_limits(xyz)

Example

lts = initialize(ThorlabsLTS150)

move_xyz(lts, 0.1, 0.1, 0.1)

# Move 0.05 meters forwards
move_x_rel(lts, 0.05)

# Get position of x stage (0.05 here)
pos_x(lts)

# Move 0.05 meters backwards
move_x_rel(lts, -0.05)

# Moves device to home position
home(lts)

# Returns x,y,z positions
pos_xyz(lts)

# First tuple contains lower limits, second contains upper limits
# (x_low_lim, y_low_lim, z_low_lim), (x_up_lim, y_up_lim, z_up_lim)
# Arrays can be used instead of tuples as well []
set_limits(lts, (0.01, 0.01, 0.01), (0.1, 0.1, 0.1))

# Will return a pair of tuples with limits you just set
get_limits(lts)

# Will return lower and upper limit for x stage
lower_x, upper_x = get_limits_x(lts)

# Will stay at 0.1 (upper limit)
move_x_abs(lts, 0.2)

# Beyond device limit but will stay at 0.1 (upper limit)
move_x_abs(lts, 5)

# Will move to 0.01 (lower limit)
move_x_abs(lts, 0)

# Clear limits
clear_limits(lts)

# Moving beyond your physical device with no limits will throw an error
# Don't do this
move_x_abs(lts, 5)

InstrumentConfig.initialize — Method

xyz = initialize(ThorlabsLTS150)

Connect to Thorlabs LTS

Returns:

ThorlabsLTS150: Device Handle

ThorlabsLTStage.get_limits — Method

Returns

(lts.x_low_limit, lts.y_low_limit, lts.z_low_limit),
   (lts.x_high_limit, lts.y_high_limit, lts.z_high_limit)

ThorlabsLTStage.get_limits_x — Method

Returns

(x_low_limit, x_high_limit)

ThorlabsLTStage.get_limits_y — Method

Returns

(y_low_limit, y_high_limit)

ThorlabsLTStage.get_limits_z — Method

Returns

(z_low_limit, z_high_limit)

ThorlabsLTStage.home — Method

home(xyz)

This will home the x, y and z stages at the same time

ThorlabsLTStage.home_x — Method

home_x(xyz)

This will home the x stage

ThorlabsLTStage.home_y — Method

home_y(xyz)

This will home the y stage

ThorlabsLTStage.home_z — Method

home_z(xyz)

This will home the z stage

ThorlabsLTStage.microsteps_per_m — Method

Returns the number of microsteps in x meters for Thorlabs LTS150

ThorlabsLTStage.microsteps_per_m — Method

Returns the number of microsteps in one meter for Thorlabs LTS150

ThorlabsLTStage.microsteps_per_mm — Method

Returns the number of microsteps in one milimeter for Thorlabs LTS150

ThorlabsLTStage.move_x_abs — Method

move_x(xyz, x)

Moves x stage to desired absolute location

ThorlabsLTStage.move_x_rel — Method

move_x(xyz, x)

Moves x stage forward or backwards.

A positive number will move it forwards along the x axis while a negative number will move it backwards.

ThorlabsLTStage.move_xyz — Method

move_xyz(xyz, x, y, z)

Simulatenously moves x, y and z stage to desired location

ThorlabsLTStage.move_y_abs — Method

move_y(xyz, y)

Moves y stage to desired absolute location

ThorlabsLTStage.move_y_rel — Method

move_y(xyz, y)

Moves y stage forward or backwards.

A positive number will move it forwards along the y axis while a negative number will move it backwards.

ThorlabsLTStage.move_z_abs — Method

move_z(xyz, z)

Moves z stage to desired absolute location

ThorlabsLTStage.move_z_rel — Method

move_z(xyz, z)

Moves z stage forward or backwards.

A positive number will move it forwards along the z axis while a negative number will move it backwards.

ThorlabsLTStage.pos_x — Method

pos_x(xyz)

Returns the current position of x stage

ThorlabsLTStage.pos_xyz — Method

pos_xyz(xyz)

Returns location of x, y and z stages in the form of a Array: [x, y, z]

ThorlabsLTStage.pos_y — Method

pos_y(xyz)

Returns the current position of y stage

ThorlabsLTStage.pos_z — Method

pos_z(xyz)

Returns the current position of z stage

ThorlabsLTStage.set_limits — Method

set_limits(xyz, (x_low_lim, y_low_lim, z_low_lim), (x_high_lim, y_high_lim, z_high_lim))

Arguments

low: A Pair or an Array of three positions: (lts.x_low_limit, lts.y_low_limit, lts.z_low_limit)

high: A Pair or an Array of three positions: (lts.x_high_limit, lts.y_high_limit, lts.z_high_limit)